fix: refactor load_config.c to comply with 42 Norm
## Changes
- Split load_config.c into two files to comply with Norm rules
- Created load_config_utils.c for utility functions
- Extracted helper functions to reduce function line count

## Details
1. Created srcs/parse/config/load_config_utils.c:
   - Moved extract_texture_path()
   - Moved parse_rgb_component()
   - Moved parse_rgb_color()

2. Refactored srcs/parse/config/load_config.c:
   - Created store_texture_path() helper function
   - Simplified store_config_value() from 38 lines to 10 lines
   - Created process_config_line() helper function
   - Reduced load_config() from 29 lines to 23 lines
   - Now has 4 functions (under 5-function limit)

3. Updated srcs/parse/config/validate_format.c:
   - Removed duplicate parse_rgb_component() function
   - Now uses shared function from load_config_utils.c

4. Updated parse.h and Makefile to include new file

## Norm Compliance
All files now pass norminette validation:
- load_config.c: OK
- load_config_utils.c: OK
- validate_format.c: OK
